requirements.txt
aiogram>=3.4,<4.0
pillow>=10.0,<11.0

